LPS-57922 Split away the tracker from the Servlet
