+++ modules/portal/portal-js-loader-modules-extender/src/com/liferay/portal/js/loader/modules/extender/JSLoaderModulesPortalWebResources.java
@@ -33,8 +33,7 @@ public class JSLoaderModulesPortalWebResources {
 	protected void activate(BundleContext bundleContext) {
 		try {
 			com.liferay.portal.kernel.servlet.PortalWebResources
-				portalWebResources = new InternalPortalWebResources(
-					_jsLoaderModulesServlet);
+				portalWebResources = new InternalPortalWebResources();
 
 			_serviceRegistration = bundleContext.registerService(
 				com.liferay.portal.kernel.servlet.PortalWebResources.class,
@@ -59,17 +58,19 @@ public class JSLoaderModulesPortalWebResources {
 		_jsLoaderModulesServlet = jsLoaderModulesServlet;
 	}
 
+	@Reference
+	protected void setJsLoaderModulesTracker(
+		JSLoaderModulesTracker jsLoaderModulesTracker) {
+
+		_jsLoaderModulesTracker = jsLoaderModulesTracker;
+	}
+
 	private JSLoaderModulesServlet _jsLoaderModulesServlet;
+	private JSLoaderModulesTracker _jsLoaderModulesTracker;
 	private ServiceRegistration<?> _serviceRegistration;
 
 	private class InternalPortalWebResources
 		implements com.liferay.portal.kernel.servlet.PortalWebResources {
-
-		public InternalPortalWebResources(
-			JSLoaderModulesServlet jsLoaderModulesServlet) {
-
-			_jsLoaderModulesServlet = jsLoaderModulesServlet;
-		}
 
 		@Override
 		public String getContextPath() {
@@ -80,7 +81,7 @@ public class JSLoaderModulesPortalWebResources {
 
 		@Override
 		public long getLastModified() {
-			return _jsLoaderModulesServlet.getTrackingCount();
+			return _jsLoaderModulesTracker.getTrackingCount();
 		}
 
 		@Override
@@ -93,8 +94,6 @@ public class JSLoaderModulesPortalWebResources {
 		public ServletContext getServletContext() {
 			return _jsLoaderModulesServlet.getServletContext();
 		}
-
-		private JSLoaderModulesServlet _jsLoaderModulesServlet;
 
 	}
